Add files via upload
In this project, the first page is the welcome page which has the logo and the features of the quiz. Also, it has an option to start the quiz.
After clicking on that button, another page will come which has questions and their options,. User can choose to answer the question in a limited time and can see the score in this page.
At the end, It shows the result and there will be an option to attempt the quiz again.